🔨 refactor: [인증] 관련 함수명 수정
