+++ src/components/Common/HamburgerMenu/index.tsx
@@ -1,12 +1,12 @@
 import { useEffect, useState } from 'react';
 import { IoMenu } from 'react-icons/io5';
-import { Link, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { useLocation } from 'react-router-dom';
 import { useAtom, useAtomValue } from 'jotai';
 import { PATH } from '@/constants/path';
 import { darkAtom } from '@/store/theme';
 import { theme } from '@/theme';
-import { useSignOutMutation } from '../../../hooks/api/useSignOutMutation';
+import { useSignOut } from '../../../hooks/api/useSignOut';
 import { isLoggedInAtom } from '../../../store/auth';
 import DarkMode from '../../DarkMode';
 import { Button, Container, Menu, MenuContainer } from './index.style';
@@ -18,9 +18,8 @@ function HamburgerMenu() {
   const darkMode = useAtomValue(darkAtom);
 
   const { pathname } = useLocation();
-  const navigate = useNavigate();
 
-  const { mutateSignOut } = useSignOutMutation();
+  const { mutateSignOut } = useSignOut();
 
   useEffect(() => {
     setToggle(false);
