+++ src/hooks/api/useUserList.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
-import { getUserList } from '@/api/member';
+import { getUserList } from '@/api/user';
 
-export const useUserListQuery = () => {
+export const useUserList = () => {
   return useQuery({
     queryKey: ['userList'],
     queryFn: getUserList
